[General] Changed the event machine interface.
After refactoring some of the code to reflect the D style guide, I was faced
with an issue. Until now, I've assumed that the event machine states are PascalCased.
However, the style guide says that enum members should be camelCased, which led me to
make a more familiar interface for switching states - 'go!"state"' replaces 'goToState'.